# backend/insomnia/strobe-openapi.yaml
openapi: 3.0.3
info:
  title: Strobe Local API
  version: 1.0.0
  contact:
    name: CAB432 Teaching Team
    email: cab432@qut.edu.au
  description: |
    Local OpenAPI specification for the Strobe server.

    This spec is designed for direct import into Insomnia or other API clients and reflects
    the request and response shapes implemented by the Express server in `server/src`.

    Authentication uses JWT bearer tokens. Register or log in first, then send:

    `Authorization: Bearer <token>`
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT token only. Do not include the "Bearer " prefix.
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000
    UserIdParam:
      in: path
      name: userId
      required: true
      description: User identifier.
      schema:
        type: string
        example: user_12345678
    PostIdParam:
      in: path
      name: postId
      required: true
      description: Post identifier.
      schema:
        type: string
        example: post_12345678
...
    FileIdParam:
      in: path
      name: fileId
      required: true
      description: File identifier returned by POST /v1/uploads/url.
      schema:
        type: string
        example: 5f9439f1
    QueryLimit:
      in: query
      name: limit
      required: false
      description: Maximum number of records to return. The server caps this at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    QueryOffset:
      in: query
      name: offset
      required: false
      description: Number of records to skip before returning results.
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Title is required
...
    PostAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake_bradford
    PostStats:
      type: object
      required: [likes, comments]
      properties:
        likes:
          type: integer
          example: 3
        comments:
          type: integer
          example: 1
    Post:
      type: object
      required:
        [
          id,
          userId,
          title,
          description,
          images,
          status,
          createdAt,
          updatedAt,
          author,
          stats,
          currentUserLiked,
        ]
      properties:
        id:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        status:
          type: string
          enum: [active, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        author:
          $ref: "#/components/schemas/PostAuthor"
        stats:
          $ref: "#/components/schemas/PostStats"
        currentUserLiked:
          type: boolean
          example: false
...
    Like:
      type: object
      required: [id, postId, userId, createdAt]
      properties:
        id:
          type: string
          example: 990e8400-e29b-41d4-a716-446655440004
        postId:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440010
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:21:00.000Z
...
    UploadedFile:
      type: object
      required: [fileId, url, size, mimetype]
      properties:
        fileId:
          type: string
          example: 5f9439f1
        url:
          type: string
          example: /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        size:
          type: integer
          example: 284193
        mimetype:
          type: string
          example: image/jpeg
...
    CreatePostRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          example: Sunrise ride, second pass
        description:
          type: string
          example: Updated after the upload finished.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
...
    UploadUrlRequest:
      type: object
      required: [postId]
      properties:
        postId:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
...
    PostResponse:
      type: object
      required: [message, post]
      properties:
        message:
          type: string
          example: Success
        post:
          $ref: "#/components/schemas/Post"
    PostListResponse:
      type: object
      required: [message, posts]
      properties:
        message:
          type: string
          example: Success
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
...
    LikeResponse:
      type: object
      required: [message, like]
      properties:
        message:
          type: string
          example: Post liked
        like:
          $ref: "#/components/schemas/Like"
...
    UploadUrlResponse:
      type: object
      required: [message, uploadUrl, fileId]
      properties:
        message:
          type: string
          example: Success
        uploadUrl:
          type: string
          example: /v1/uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1
        fileId:
          type: string
          example: 5f9439f1
    UploadFileResponse:
      type: object
      required: [message, file]
      properties:
        message:
          type: string
          example: Success
        file:
          $ref: "#/components/schemas/UploadedFile"
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Deleted successfully
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ValidationError
            message: Title is required
    Unauthorised:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorised
            message: Invalid or expired token
    Forbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Forbidden
            message: You can only modify your own posts
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NotFound
            message: Post not found
paths:
...
  /v1/posts:
    post:
      tags: [Posts]
      operationId: createPost
      summary: Create post
      description: Creates a new post for the authenticated user. Description and images are optional.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            example:
              title: Sunrise ride
              description: Cold start, good coffee, and an empty river path.
              images:
                - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
      responses:
        "201":
          description: Post created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              example:
                message: Created successfully
                post:
                  id: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  title: Sunrise ride
                  description: Cold start, good coffee, and an empty river path.
                  images:
                    - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
                  status: active
                  createdAt: 2026-07-08T09:10:00.000Z
                  updatedAt: 2026-07-08T09:10:00.000Z
                  hiddenBy: null
                  hiddenAt: null
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake_bradford
                  stats:
                    likes: 0
                    comments: 0
                  currentUserLiked: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
  /v1/posts/user/{userId}:
    get:
      tags: [Posts]
      operationId: getUserPosts
      summary: Get posts by user
      description: Returns the target user's posts in reverse chronological order. Hidden posts are only visible to the author and moderators.
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: User posts returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
              example:
                message: Success
                posts:
                  - id: 660e8400-e29b-41d4-a716-446655440001
                    userId: 550e8400-e29b-41d4-a716-446655440000
                    title: Sunrise ride
                    description: Cold start, good coffee, and an empty river path.
                    images: []
                    status: active
                    createdAt: 2026-07-08T09:10:00.000Z
                    updatedAt: 2026-07-08T09:10:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      username: jake_bradford
                    stats:
                      likes: 3
                      comments: 1
                    currentUserLiked: false
        "404":
          description: Target user does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
  /v1/posts/{id}:
    get:
      tags: [Posts]
      operationId: getPost
      summary: Get post
      description: Retrieves one post. Hidden posts are only visible to the post owner and moderators.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Post returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              example:
                message: Success
                post:
                  id: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  title: Sunrise ride
                  description: Cold start, good coffee, and an empty river path.
                  images:
                    - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
                  status: active
                  createdAt: 2026-07-08T09:10:00.000Z
                  updatedAt: 2026-07-08T09:10:00.000Z
                  hiddenBy: null
                  hiddenAt: null
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake_bradford
                  stats:
                    likes: 3
                    comments: 1
                  currentUserLiked: false
        "403":
          description: Hidden post requested by a non-owner, non-moderator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You do not have permission to view this post
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Posts]
      operationId: updatePost
      summary: Update post
      description: Updates a post. Only the post author can perform this action.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
            example:
              title: Sunrise ride, second pass
              description: Updated after the upload finished.
              images:
                - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
      responses:
        "200":
          description: Post updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              example:
                message: Updated successfully
                post:
                  id: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  title: Sunrise ride, second pass
                  description: Updated after the upload finished.
                  images:
                    - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
                  status: active
                  createdAt: 2026-07-08T09:10:00.000Z
                  updatedAt: 2026-07-08T09:22:00.000Z
                  hiddenBy: null
                  hiddenAt: null
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake_bradford
                  stats:
                    likes: 3
                    comments: 1
                  currentUserLiked: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete post
      description: Deletes a post and cascades associated comments and likes. Only the post author can delete it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Post deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Non-owner attempted delete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You can only delete your own posts
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/posts/{id}/like:
    post:
      tags: [Posts]
      operationId: likePost
      summary: Like post
      description: Adds a like from the authenticated user to the target post.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "201":
          description: Like created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
              example:
                message: Post liked
                like:
                  id: 990e8400-e29b-41d4-a716-446655440004
                  postId: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440010
                  createdAt: 2026-07-08T09:21:00.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: User already liked the post.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Conflict
                message: You have already liked this post
    delete:
      tags: [Posts]
      operationId: unlikePost
      summary: Unlike post
      description: Removes the authenticated user's like from the target post.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Like removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Post unliked
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Post not found or user had not liked the post.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                postMissing:
                  value:
                    error: NotFound
                    message: Post not found
                likeMissing:
                  value:
                    error: NotFound
                    message: You have not liked this post
  /v1/posts/{id}/hide:
    post:
      tags: [Posts]
      operationId: hidePost
      summary: Hide post
      description: Moderator-only action that marks a post as hidden.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Post hidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              example:
                message: Updated successfully
                post:
                  id: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  title: Sunrise ride
                  description: Cold start, good coffee, and an empty river path.
                  images: []
                  status: hidden
                  createdAt: 2026-07-08T09:10:00.000Z
                  updatedAt: 2026-07-08T09:40:00.000Z
                  hiddenBy: 550e8400-e29b-41d4-a716-446655440999
                  hiddenAt: 2026-07-08T09:40:00.000Z
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: jackson_riding
                  stats:
                    likes: 3
                    comments: 1
                  currentUserLiked: false
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Caller is authenticated but not a moderator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: Moderator role required
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/uploads/url:
    post:
      tags: [Uploads]
      operationId: getUploadUrl
      summary: Get upload URL
      description: |
        Generates a one-time local upload path for an authenticated user's post image.
        The returned uploadUrl should then be used with PUT multipart form-data.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
            example:
              postId: 660e8400-e29b-41d4-a716-446655440001
      responses:
        "200":
          description: Upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
              example:
                message: Success
                uploadUrl: /v1/uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1
                fileId: 5f9439f1
        "400":
          description: postId missing from request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: ValidationError
                message: postId is required to generate an upload URL
        "401":
          $ref: "#/components/responses/Unauthorised"
  /v1/uploads/{userId}/{postId}/{fileId}:
    put:
      tags: [Uploads]
      operationId: uploadFile
      summary: Upload file
      description: |
        Uploads a single image using multipart form-data.

        Requirements:
        - Field name must be file
        - Allowed MIME types: image/jpeg, image/png, image/gif, image/webp
        - Maximum file size: 10 MB
        - userId in the path must match the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/FileIdParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File stored locally and public URL returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadFileResponse"
              example:
                message: Success
                file:
                  fileId: 5f9439f1
                  url: /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
                  size: 284193
                  mimetype: image/jpeg
        "400":
          description: Missing file or invalid file payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingFile:
                  value:
                    error: ValidationError
                    message: No file provided
                invalidMime:
                  value:
                    error: ValidationError
                    message: "Invalid file type. Allowed: JPEG, PNG, GIF, WebP"
                fileTooLarge:
                  value:
                    error: ValidationError
                    message: "File too large. Maximum size: 10MB"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: userId in the path does not match the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You can only upload files for your own user ID